+++ test/unit/fuzz-tools.test.js
@@ -332,6 +332,7 @@ import * as lanskyV352 from '../../lib/lansky-v352.js';
 import * as croweDdhV353 from '../../lib/crowe-ddh-v353.js';
 import * as tonnisHipOaV354 from '../../lib/tonnis-hip-oa-v354.js';
 import * as lachmanAclV355 from '../../lib/lachman-acl-v355.js';
+import * as ceapVenousV356 from '../../lib/ceap-venous-v356.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -609,6 +610,7 @@ const MODULES = {
   'crowe-ddh-v353.js': croweDdhV353,
   'tonnis-hip-oa-v354.js': tonnisHipOaV354,
   'lachman-acl-v355.js': lachmanAclV355,
+  'ceap-venous-v356.js': ceapVenousV356,
   'tb-testing.js': tbTesting,
 };